backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.1.1
selenium==4.9.1
SQLAlchemy==2.0.19
python-dotenv
Gunicorn